requirements.txt
feedparser==6.0.11
requests==2.31.0
beautifulsoup4==4.12.3
openai==1.84.0
resend==0.8.0
schedule==1.2.0
python-dotenv==1.0.1